+++ packages/sdk/stories/helpers/sync_source_options.ts
@@ -1,0 +1,12 @@
+export default [
+    {
+        id: 'a',
+        name: 'EMEA Sales Tracker',
+        color: 'yellowLight2',
+    },
+    {
+        id: 'b',
+        name: 'Salesforce - LATAM Top-Level Report',
+        color: 'orangeLight2',
+    },
+];